hsvtt: Add eslint support; fix findings
This patch adds eslint support to the project. That includes an eslintrc file
that is based on the Airbnb standard, as well as an eslintignore file. This
patch opts for an ES5 'legacy' standard, as the lack of a transpiler for the
front-end Javascript code prevents using ES6.

A few warnings and errors have been left unresolved. These include:
* A single console statement indicating that the app is running. This should
  be replaced with a logger, such a bunyan.

* A few references to undefined variables. The intent of this function could
  not be determined, nor how it functioned.

* An unused variable 'orderedStops', which appears to either have no purpose
  or should be the variable that is returned from a particular function.

It's important to note that despite the scope of the changes, no semantic
changes should have been introduced in this patch. However, given the scope of
the changes, significant testing should be done before deploying the results
of this patch into a production environment.